rename and remove old models/interfaces
