+++ pages/api/report/stop.ts
@@ -2,7 +2,8 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { PutCommand, DynamoDBDocumentClient } from "@aws-sdk/lib-dynamodb";
 
-import { Stop } from "@/models/reports/stop";
+import { Report } from "@/models/report";
+import { ReportType } from "@/enums/reportType";
 
 const client = new DynamoDBClient({
   region: process.env.AWS_REGION,
@@ -14,14 +15,20 @@ export default async function handler(
   res: NextApiResponse
 ) {
   if (req.method === "POST") {
-    const stopReport = new Stop(req.body.stopID, req.body.personCount);
+    const report = new Report({
+      stop_id: req.body.stop_id, 
+      report_type: ReportType.StopCrowded,
+      bus_id: undefined,
+      route_id: undefined,
+      person_count: req.body.person_count
+    });
 
     const command = new PutCommand({
-      TableName: process.env.AWS_STOP_TABLE_NAME,
-      Item: stopReport,
+      TableName: process.env.AWS_TABLE_NAME,
+      Item: report,
     });
 
     await docClient.send(command);
-    res.status(201).json(stopReport);
+    res.status(201).json(report);
   }
 }
